Script to fetch indexable content from search
Fetches the indexable content of GOV.UK pages using the search API, and
prints it as CSV to stdout.

Indexable content is format-dependent: it's whatever publishing
applications decide to pass to rummager when indexing the document.

This leaves some empty rows where data can't be found - some links are
not indexed in search so we can't fetch their content.

I think the missing ones fall into one of three cases:

- Special document types that we've never made searchable, such as the
  about pages of topical events (these are by definition no longer
  relevant)
  Example: https://www.gov.uk/government/topical-events/national-apprenticeship-week-2016/about

- Pages that were published in error. These display a special page
  containing a link to the replacement content. It's easier to fix these
ones in the source CSV.
  Example: https://www.gov.uk/government/news/new-framework-for-inspection-of-childrens-services

- Pages that have been completely removed from GOV.UK without leaving a
  replacement (410 Gone pages)