Treat stopwords as UTF8 bytestrings
The NLTK stopwords are `unicode` strings, whereas the `lemmatize`
function always returns `str` (in this case, utf8 encoded bytes).
https://docs.python.org/2/howto/unicode.html#python-2-x-s-unicode-support

Mixing these up causes problems because to convert between them
python needs to assume an encoding, and the default encoding is ascii.

When we check if a bigram is in the list of stopwords, python never
accepts any with non-ascii characters in, eg `'interesting/JJ_\xc3\xa9mile/NN'`

By default, python with log this as a warning, such as:

```
>>> 'é' in [u'é']
__main__:1: UnicodeWarning: Unicode equal comparison failed to convert
both arguments to Unicode - interpreting them as being unequal
False
```

I've added the `filterwarnings` call to make this an error, and
converted all the stop words to `str`s.

This shouldn't change the output right now, because we currently
whitelist the bigrams.