CLI option to tag a different set of documents
Making this an option allows us to remove a few scripts with very
similar code.

I've included headers in the source data since some of them have
slightly different shapes (some have the source audit as a column).

Early years curated used to tag a different document set to the one it
was trained on, but we don't actually need this functionality at the
moment.

Dictionaries can now be saved and reused in a later run of the script.
At the moment, for this to work properly, I think the second run needs
to use the same corpus as the original run (i.e. the `self.lemmas` list
in gensim engine needs to stay the same).