Save experiment model, corpus, and dictionary
Each time we run an experiment, save all the state, so we
can load it again later to generate visualisations, topics, tags.
Filenames are generated based on date.

This should help us iterate faster when we need to change things
that happen later in the pipeline.

The `Experiment` class encapsulates the trained LDA model for a corpus.
GensimEngine now returns an instance of this, which should get rid
of some of the time-dependence between engine attributes.

I haven't moved all of the output methods to `Experiment` yet;
I've just focused on visualisation for now. I think some of the
tagging code could be simplified to reuse the existing corpus
data, rather than redoing the lemmatization.