Rework the phrasemachine experiment
We detected a bug in phrasemachine, which meant the previous results
weren't correct.

This commit adds the results of running the same experiment with the
fixed version of phrasemachine. Both the model and the model state files
have been compressed in order to be able to push them to GitHub. Please
decompress them in case you wish to refine the experiment.

Note: we should bump phrasemachine's version once that's been released.

Phrasemachine's PR:
https://github.com/slanglab/phrasemachine/commit/1813930786b4ca4e2898c87640ba47a6cc1b112f